backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
requests>=2.31
django-filter>=23.5
drf-spectacular>=0.27
django-storages>=1.14
boto3>=1.34
